+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @razorpay/blade
 
+## 12.120.0
+
+### Minor Changes
+
+- 9edbd8f33: update design tokens from Figma
+
+  Added 32 tokens and removed 0 tokens.
+
+## 12.119.1
+
+### Patch Changes
+
+- 83be08d99: feat(tokens): add new tokens - `onNeutral` has been added to `interactive.text` & `interactive.icon`
+
 ## 12.119.0
 
 ### Minor Changes